+++ contracts/mockContracts/PublicSortedLinkedList.sol
@@ -4,15 +4,15 @@ import "../SortedLinkedList.sol";
 
 contract PublicSortedLinkedList is SortedLinkedList {
 
-    function getOrderDetails_p(uint64 orderId)
+    function getOrderDetails_p(uint32 orderId)
         public
         view
         returns (
             address _maker,
             uint128 _srcAmount,
             uint128 _dstAmount,
-            uint64 _prevId,
-            uint64 _nextId
+            uint32 _prevId,
+            uint32 _nextId
         )
     {
         return super.getOrderDetails(orderId);
@@ -20,37 +20,37 @@ contract PublicSortedLinkedList is SortedLinkedList {
 
     function add_p(uint128 srcAmount, uint128 dstAmount)
         public
-        returns(uint64)
+        returns(uint32)
     {
         return super.add(srcAmount, dstAmount);
     }
 
-    function addAfterId_p(uint128 srcAmount, uint128 dstAmount, uint64 prevId)
+    function addAfterId_p(uint128 srcAmount, uint128 dstAmount, uint32 prevId)
         public
-        returns(uint64)
+        returns(uint32)
     {
         return super.addAfterId(srcAmount, dstAmount, prevId);
     }
 
-    function removeById_p(uint64 orderId) public {
+    function removeById_p(uint32 orderId) public {
         return super.removeById(orderId);
     }
 
-    function update_p(uint64 orderId, uint128 srcAmount, uint128 dstAmount)
+    function update_p(uint32 orderId, uint128 srcAmount, uint128 dstAmount)
         public
-        returns(uint64)
+        returns(uint32)
     {
         return super.update(orderId, srcAmount, dstAmount);
     }
 
     function updateWithPositionHint_p(
-        uint64 orderId,
+        uint32 orderId,
         uint128 srcAmount,
         uint128 dstAmount,
-        uint64 prevId
+        uint32 prevId
     )
         public
-        returns(uint64)
+        returns(uint32)
     {
         return super.updateWithPositionHint(
             orderId,
